requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
plotly==5.23.0
slowapi==0.1.9
boto3
